+++ lib/imports/constants_and_requires.rb
@@ -7,7 +7,6 @@ require 'rest-client'
 require 'timeout'
 require 'uri'
 require 'fileutils'
-require 'colored'
 require 'yaml'
 require 'date'
 require 'optparse'
@@ -22,6 +21,7 @@ require_relative '../../lib/misc/legal'
 require_relative '../../lib/misc/spider'
 require_relative '../../lib/modules/copy'
 require_relative '../../lib/modules/site_info'
+require_relative '../../lib/modules/expansion/string_expan'
 
 # Modules that need to be included
 include Format
@@ -46,9 +46,19 @@ USER_AGENTS = { # Temporary fix for user agents until I can refactor the YAML fi
     10 => 'Mozilla/5.0 (compatible; bingbot/2.0; +http://www.bing.com/bingbot.htm)',
 }
 =end
+FORMAT = Format::StringFormat.new
 PATH = Dir.pwd
 VERSION = Whitewidow.version
 SEARCH = File.readlines("#{PATH}/lib/lists/search_query.txt").sample
-USER_AGENTS = YAML.load_file("#{PATH}/lib/lists/rand-age.yml")
+agents = YAML.load_file("#{PATH}/lib/lists/rand-age.yml")
 OPTIONS = {}
-USER_AGENT = USER_AGENTS[rand(1..10)]
+USER_AGENT = agents[rand(1..20)]
+SKIP = %w(/webcache.googleusercontent.com stackoverflow.com github.com)
+LOADING_ERRORS = [RestClient::ResourceNotFound, RestClient::InternalServerError, RestClient::RequestTimeout,
+                RestClient::Gone, RestClient::SSLCertificateNotVerified, RestClient::Forbidden,
+                OpenSSL::SSL::SSLError, Errno::ECONNREFUSED, URI::InvalidURIError, Errno::ECONNRESET,
+                Timeout::Error, OpenSSL::SSL::SSLError, Zlib::GzipFile::Error, RestClient::MultipleChoices,
+                RestClient::Unauthorized, SocketError, RestClient::BadRequest, RestClient::ServerBrokeConnection,
+                RestClient::MaxRedirectsReached]
+FATAL_ERRORS = [Mechanize::ResponseCodeError, RestClient::ServiceUnavailable, OpenSSL::SSL::SSLError,
+                RestClient::BadGateway]