+++ lib/modules/format.rb
@@ -1,53 +1,70 @@
 # Color format the output of the program based on severity of the issue
+require_relative "../../lib/modules/expansion/string_expan"
+
 module Format
 
-  # @param [String] *args
-  # Yellow is a warning
-  def warning(input)
-    puts "[#{Time.now.strftime("%T")} WARNING] #{input}".yellow.bold
-  end
+  class StringFormat
 
-  # Red is a bad thing, this particular instance will kill the program
-  def fatal(input)
-    puts "[#{Time.now.strftime("%T")} FATAL] #{input}".red.bold
-  end
+    # @param [String] *args
+    # Yellow is a warning
+    def warning(input)
+      puts "[#{Time.now.strftime("%T").purple} WARNING] #{input.yellow}"
+    end
 
-  # Red is still bad
-  def err(input)
-    puts "[#{Time.now.strftime("%T")} ERROR] #{input}".red.bold
-  end
+    # Red is a bad thing, this particular instance will kill the program
+    def fatal(input)
+      puts "[#{Time.now.strftime("%T").purple} FATAL] #{input.red.bold}"
+    end
 
-  # Green is good
-  def info(input)
-    puts "[#{Time.now.strftime("%T")} INFO] #{input}".green.bold
-  end
+    # Red is still bad
+    def err(input)
+      puts "[#{Time.now.strftime("%T").purple} ERROR] #{input.red}"
+    end
 
-  # Found a site wooohooo
-  def success(input)
-    puts "[#{Time.now.strftime("%T")} SUCCESS] #{input}".white.bold
-  end
+    # Green is good
+    def info(input)
+      puts "[#{Time.now.strftime("%T").purple} INFO] #{input.green.bold}"
+    end
 
-  # Throws the valuable information that you will need, this is very good.
-  def site_found(input)
+    # Found a site wooohooo
+    def success(input)
+      puts "[#{Time.now.strftime("%T").purple} SUCCESS] #{input.white}"
+    end
 
-    success = 'Possible Vulnerability'.green.bold
-    puts <<-_END_
+    # Throws the valuable information that you will need, this is very good.
+    def site_found(input)
+
+      success = 'Possible Vulnerability'
+      puts <<-_END_
     [ x ] -----------------------------------------------
 
-    [#{Time.now.strftime("%T").green.bold} #{success}]
+    [#{Time.now.strftime("%T")} #{success}]
 
-    [ X ]    Injection Point       :: #{input.cyan.bold}
-    [ X ]    Server Specifications :: #{SiteInfo.capture_host(input).cyan.bold}
-    [ X ]    Possible IP Address   :: #{SiteInfo.capture_ip(input).cyan.bold}
+    [ X ]    Injection Point       :: #{input}
+    [ X ]    Server Specifications :: #{SiteInfo.capture_host(input)}
+    [ X ]    Possible IP Address   :: #{SiteInfo.capture_ip(input)}
 
-            _END_
-  end
+      _END_
+    end
 
-  # And a normal usage page
-  def usage(input)
-    puts
-    puts "[USAGE]#{input}".white.bold
-    puts
+    # And a normal usage page
+    def usage(input)
+      puts
+      puts "[USAGE]#{input}"
+      puts
+    end
+
   end
 
 end
+
+=begin
+##
+# Testing
+#
+test = Format::StringFormat.new
+test.err("test")
+test.info("test")
+test.warning("test")
+test.fatal("test")
+=end