+++ whitewidow.rb
@@ -9,7 +9,7 @@ require_relative 'lib/imports/constants_and_requires'
 # Usage page, shows basic shell of commands
 #
 def usage_page
-  Format.usage("You can run me with the following flags: #{File.basename(__FILE__)} -[d|e|h] -[f] <path/to/file/if/any>")
+  FORMAT.usage("You can run me with the following flags: #{File.basename(__FILE__)} -[d|e|h] -[f] <path/to/file/if/any>")
   exit
 end
 
@@ -17,11 +17,11 @@ end
 # Examples page, gives info on how to use the program
 #
 def examples_page
-  Format.usage('This is my examples page, I\'ll show you a few examples of how to get me to do what you want.')
-  Format.usage('Running me with a file: whitewidow.rb -f <path/to/file> keep the file inside of one of my directories.')
-  Format.usage('Running me default, if you don\'t want to use a file, because you don\'t think I can handle it, or for whatever reason, you can run me default by passing the Default flag: whitewidow.rb -d this will allow me to scrape Google for some SQL vuln sites, no guarentees though!')
-  Format.usage('Running me with my Help flag will show you all options an explanation of what they do and how to use them')
-  Format.usage('Running me without a flag will show you the usage page. Not descriptive at all but gets the point across')
+  FORMAT.usage('This is my examples page, I\'ll show you a few examples of how to get me to do what you want.')
+  FORMAT.usage('Running me with a file: whitewidow.rb -f <path/to/file> keep the file inside of one of my directories.')
+  FORMAT.usage('Running me default, if you don\'t want to use a file, because you don\'t think I can handle it, or for whatever reason, you can run me default by passing the Default flag: whitewidow.rb -d this will allow me to scrape Google for some SQL vuln sites, no guarentees though!')
+  FORMAT.usage('Running me with my Help flag will show you all options an explanation of what they do and how to use them')
+  FORMAT.usage('Running me without a flag will show you the usage page. Not descriptive at all but gets the point across')
 end
 
 #
@@ -56,7 +56,7 @@ end
 # File formatting
 #
 def format_file
-  Format.info('Writing to temporary file..')
+  FORMAT.info('Writing to temporary file..')
   if File.exists?(OPTIONS[:file])
     file = Tempfile.new('file')
     IO.read(OPTIONS[:file]).each_line do |s|
@@ -66,7 +66,7 @@ def format_file
       File.open("#{PATH}/tmp/#sites.txt", 'a+') { |line| line.puts(file) }
     end
     file.unlink
-    Format.info("File: #{OPTIONS[:file]}, has been formatted and saved as #sites.txt in the tmp directory.")
+    FORMAT.info("File: #{OPTIONS[:file]}, has been formatted and saved as #sites.txt in the tmp directory.")
   else
     puts <<-_END_
 
@@ -86,8 +86,8 @@ end
 # Get the URLS by connecting to google and scraping for the URLS on the first page
 #
 def get_urls
-  Format.info("I'll run in default mode!")
-  Format.info("I'm searching for possible SQL vulnerable sites, using search query #{SEARCH}")
+  FORMAT.info("I'll run in default mode!")
+  FORMAT.info("I'm searching for possible SQL vulnerable sites, using search query #{SEARCH}")
   agent = Mechanize.new
   agent.user_agent = USER_AGENT
   page = agent.get('http://www.google.com/')
@@ -99,16 +99,19 @@ def get_urls
       str = link.href.to_s
       str_list = str.split(%r{=|&})
       urls = str_list[1]
+      SKIP.each do |blacklist|
+        next if urls.to_s.include?(blacklist)
+      end
       #check_urls_for_blacklist(urls)
       urls_to_log = URI.decode(urls)
-      Format.success("Site found: #{urls_to_log}")
+      FORMAT.success("Site found: #{urls_to_log}")
       sleep(1)
       %w(' ` -- ;).each { |sql|
         File.open("#{PATH}/tmp/SQL_sites_to_check.txt", 'a+') { |s| s.puts("#{urls_to_log}#{sql}") }
       }
     end
   end
-  Format.info("I've dumped possible vulnerable sites into #{PATH}/tmp/SQL_sites_to_check.txt")
+  FORMAT.info("I've dumped possible vulnerable sites into #{PATH}/tmp/SQL_sites_to_check.txt")
 end
 =begin # trying to fix the google cache literally made it worse lol
 def check_urls_for_blacklist(url)
@@ -127,34 +130,34 @@ def vulnerability_check
   when OPTIONS[:default]
     file_to_read = "tmp/SQL_sites_to_check.txt"
   when OPTIONS[:file]
-    Format.info("Let's check out this file real quick like..")
+    FORMAT.info("Let's check out this file real quick like..")
     file_to_read = "tmp/#sites.txt"
   end
-  Format.info('Forcing encoding to UTF-8') unless OPTIONS[:file]
+  FORMAT.info('Forcing encoding to UTF-8') unless OPTIONS[:file]
   IO.read("#{PATH}/#{file_to_read}").each_line do |vuln|
     begin
-      Format.info("Parsing page for SQL syntax error: #{vuln.chomp}")
+      FORMAT.info("Parsing page for SQL syntax error: #{vuln.chomp}")
       Timeout::timeout(10) do
         vulns = vuln.encode(Encoding.find('UTF-8'), {invalid: :replace, undef: :replace, replace: ''}) # Force encoding to UTF-8
         begin
           if parse("#{vulns.chomp}'", 'html', 0)[/You have an error in your SQL syntax/] # IS this really what I'm relying on?
-            Format.site_found(vulns.chomp)
+            FORMAT.site_found(vulns.chomp)
             File.open("#{PATH}/tmp/SQL_VULN.txt", "a+") { |s| s.puts(vulns) }
             sleep(1)
           else
-            Format.warning("URL: #{vulns.chomp} is not vulnerable, dumped to non_exploitable.txt")
+            FORMAT.warning("URL: #{vulns.chomp} is not vulnerable, dumped to non_exploitable.txt")
             File.open("#{PATH}/log/non_exploitable.txt", "a+") { |s| s.puts(vulns) }
             sleep(1)
           end
         rescue Timeout::Error, OpenSSL::SSL::SSLError
-          Format.warning("URL: #{vulns.chomp} failed to load dumped to non_exploitable.txt")
+          FORMAT.warning("URL: #{vulns.chomp} failed to load dumped to non_exploitable.txt")
           File.open("#{PATH}/log/non_exploitable.txt", "a+") { |s| s.puts(vulns) }
           sleep(1)
           next
         end
       end
-    rescue RestClient::ResourceNotFound, RestClient::InternalServerError, RestClient::RequestTimeout, RestClient::Gone, RestClient::SSLCertificateNotVerified, RestClient::Forbidden, OpenSSL::SSL::SSLError, Errno::ECONNREFUSED, URI::InvalidURIError, Errno::ECONNRESET, Timeout::Error, OpenSSL::SSL::SSLError, Zlib::GzipFile::Error, RestClient::MultipleChoices, RestClient::Unauthorized, SocketError, RestClient::BadRequest, RestClient::ServerBrokeConnection, RestClient::MaxRedirectsReached => e
-      Format.err("URL: #{vuln.chomp} failed due to an error while connecting, URL dumped to non_exploitable.txt") # I'll be putting the error also
+    rescue *LOADING_ERRORS
+      FORMAT.err("URL: #{vuln.chomp} failed due to an error while connecting, URL dumped to non_exploitable.txt") # I'll be putting the error also
       File.open("#{PATH}/log/non_exploitable.txt", "a+") { |s| s.puts(vuln) }
       next
     end
@@ -174,18 +177,18 @@ case
       Legal.legal
       get_urls
       vulnerability_check unless File.size("#{PATH}/tmp/SQL_sites_to_check.txt") == 0
-      Format.warning("No sites found for search query: #{SEARCH}. Logging into error_log.LOG. Create a issue regarding this.") if File.size("#{PATH}/tmp/SQL_sites_to_check.txt") == 0
-      File.open("#{PATH}/log/error_log.LOG", 'a+') { |s| s.puts("No sites found with search querie #{SEARCH}") } if File.size("#{PATH}/tmp/SQL_sites_to_check.txt") == 0
+      FORMAT.warning("No sites found for search query: #{SEARCH}. Logging into error_log.LOG. Create a issue regarding this.") if File.size("#{PATH}/tmp/SQL_sites_to_check.txt") == 0
+      File.open("#{PATH}/log/error_log.LOG", 'a+') { |s| s.puts("No sites found with search query #{SEARCH}") } if File.size("#{PATH}/tmp/SQL_sites_to_check.txt") == 0
       File.truncate("#{PATH}/tmp/SQL_sites_to_check.txt", 0)
-      Format.info("I'm truncating SQL_sites_to_check file back to #{File.size("#{PATH}/tmp/SQL_sites_to_check.txt")}")
+      FORMAT.info("I'm truncating SQL_sites_to_check file back to #{File.size("#{PATH}/tmp/SQL_sites_to_check.txt")}")
       Copy.file("#{PATH}/tmp/SQL_VULN.txt", "#{PATH}/log/SQL_VULN.LOG")
       File.truncate("#{PATH}/tmp/SQL_VULN.txt", 0)
-      Format.info("I've run all my tests and queries, and logged all important information into #{PATH}/log/SQL_VULN.LOG")
-    rescue Mechanize::ResponseCodeError, RestClient::ServiceUnavailable, OpenSSL::SSL::SSLError, RestClient::BadGateway => e
+      FORMAT.info("I've run all my tests and queries, and logged all important information into #{PATH}/log/SQL_VULN.LOG")
+    rescue *FATAL_ERRORS => e
       d = DateTime.now
-      Format.fatal("Well this is pretty crappy.. I seem to have encountered a #{e} error. I'm gonna take the safe road and quit scanning before I break something. You can either try again, or manually delete the URL that caused the error.")
+      FORMAT.fatal("Well this is pretty crappy.. I seem to have encountered a #{e} error. I'm gonna take the safe road and quit scanning before I break something. You can either try again, or manually delete the URL that caused the error.")
       File.open("#{PATH}/log/error_log.LOG", 'a+'){ |error| error.puts("[#{d.month}-#{d.day}-#{d.year} :: #{Time.now.strftime("%T")}]#{e}") }
-      Format.info("I'll log the error inside of #{PATH}/log/error_log.LOG for further analysis.")
+      FORMAT.info("I'll log the error inside of #{PATH}/log/error_log.LOG for further analysis.")
     end
   when OPTIONS[:file]
     begin
@@ -194,23 +197,23 @@ case
       Credits.credits
       sleep(1)
       Legal.legal
-      Format.info('Formatting file')
+      FORMAT.info('Formatting file')
       format_file
       vulnerability_check
       File.truncate("#{PATH}/tmp/SQL_sites_to_check.txt", 0)
-      Format.info("I'm truncating SQL_sites_to_check file back to #{File.size("#{PATH}/tmp/SQL_sites_to_check.txt")}")
+      FORMAT.info("I'm truncating SQL_sites_to_check file back to #{File.size("#{PATH}/tmp/SQL_sites_to_check.txt")}")
       Copy.file("#{PATH}/tmp/SQL_VULN.txt", "#{PATH}/log/SQL_VULN.LOG")
       File.truncate("#{PATH}/tmp/SQL_VULN.txt", 0)
-      Format.info("I've run all my tests and queries, and logged all important information into #{PATH}/log/SQL_VULN.LOG") unless File.size("#{PATH}/log/SQL_VULN.LOG") == 0
-    rescue Mechanize::ResponseCodeError, RestClient::ServiceUnavailable, OpenSSL::SSL::SSLError, RestClient::BadGateway => e
+      FORMAT.info("I've run all my tests and queries, and logged all important information into #{PATH}/log/SQL_VULN.LOG") unless File.size("#{PATH}/log/SQL_VULN.LOG") == 0
+    rescue *FATAL_ERRORS => e
       d = DateTime.now
-      Format.fatal("Well this is pretty crappy.. I seem to have encountered a #{e} error. I'm gonna take the safe road and quit scanning before I break something. You can either try again, or manually delete the URL that caused the error.")
+      FORMAT.fatal("Well this is pretty crappy.. I seem to have encountered a #{e} error. I'm gonna take the safe road and quit scanning before I break something. You can either try again, or manually delete the URL that caused the error.")
       File.open("#{PATH}/log/error_log.LOG", 'a+'){ |error| error.puts("[#{d.month}-#{d.day}-#{d.year} :: #{Time.now.strftime("%T")}]#{e}") }
-      Format.info("I'll log the error inside of #{PATH}/log/error_log.LOG for further analysis.")
+      FORMAT.info("I'll log the error inside of #{PATH}/log/error_log.LOG for further analysis.")
     end
   when OPTIONS[:example]
     examples_page
   else
-    Format.warning('You failed to pass me a flag!')
+    FORMAT.warning('You failed to pass me a flag!')
     usage_page
 end